+++ src/components/homepage/StatisticsBand/StatisticsBand.styles.ts
