+++ src/components/homepage/StatisticsBand/StatisticsBand.tsx
@@ -5,7 +5,7 @@ import Typography from "@mui/material/Typography";
 import Image from "next/image";
 import React from "react";
 import styles from "./StatisticsBand.styles";
-import livingRoomImg from "../../images/living-room.jpg";
+import livingRoomImg from "../../../images/living-room.jpg";
 
 export default function StatisticsBand() {
   return (
