+++ src/components/homepage/StatisticsBand/index.ts
