+++ src/components/homepage/index.ts
@@ -1,0 +1,4 @@
+export { default as ContactBand } from "./ContactBand";
+export { default as Hero } from "./Hero";
+export { default as ServicesBand } from "./ServicesBand";
+export { default as StatisticsBand } from "./StatisticsBand";
