+++ src/components/pattern/GlobalFooter/GlobalFooter.styles.ts
