+++ src/components/pattern/GlobalFooter/GlobalFooter.tsx
