+++ src/components/pattern/GlobalFooter/index.ts
