+++ src/components/pattern/GlobalNav/GlobalNav.styles.ts
